test: Split integration tests into method and variables
